pyproject.toml
[project]
name = "scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "pydantic>=2.10.6",
    "selenium>=4.29.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.1",
]


[tool.ruff]
line-length = 79

[tool.ruff.lint]
ignore = ["E722"]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "SIM",# flake8-simplify
    "I",  # isort
]
